SubscriptionService/pyproject.toml
[project]
name = "subscriptionservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alos", email = "daryoosh.safe@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "pydantic-settings>=2.14.0",
    "sqlalchemy>=2.0.49",
    "stripe>=15.1.0",
    "uvicorn>=0.46.0",
]

[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
